+++ src/signature-help.jl
@@ -586,12 +586,10 @@ function handle_SignatureHelpRequest(
     state = server.state
     uri = msg.params.textDocument.uri
     result = get_file_info(state, uri, cancel_flag)
-    if result isa ResponseError
-        return send(server,
-            SignatureHelpResponse(;
-                id = msg.id,
-                result = nothing,
-                error = result))
+    if isnothing(result)
+        return send(server, SignatureHelpResponse(; id = msg.id, result = null))
+    elseif result isa ResponseError
+        return send(server, SignatureHelpResponse(; id = msg.id, result = nothing, error = result))
     end
     fi = result
     pos = adjust_position(state, uri, msg.params.position)